Economy-Project/src/bot/cogs/gambling.py
from discord.ext import commands
import random

class Gambling(commands.Cog):
    """Comandos relacionados con el juego de azar"""

    def __init__(self, bot):
        self.bot = bot

    @commands.command(name="bet", aliases=["apostar"])
    async def bet(self, ctx, amount: int):
        """💸 Apuesta una cantidad de dinero en un juego de azar"""
        user_id = ctx.author.id
        modalidad = "default"  # Cambiar según la modalidad del servidor

        # Aquí deberías obtener los datos del usuario desde tu sistema de economía
        user_data = await self.get_user_data(modalidad, user_id)

        if amount <= 0:
            await ctx.send("¡Debes apostar una cantidad positiva!")
            return

        if user_data.wallet < amount:
            await ctx.send("¡No tienes suficiente dinero para apostar esa cantidad!")
            return

        # Simulación de un juego de azar (50% de ganar)
        if random.random() < 0.5:
            winnings = amount * 2
            user_data.wallet += winnings
            await self.update_user_data(modalidad, user_data)
            await ctx.send(f"¡Ganaste! Tu ganancia es de ${winnings:,}. Tu nuevo saldo es ${user_data.wallet:,}.")
        else:
            user_data.wallet -= amount
            await self.update_user_data(modalidad, user_data)
            await ctx.send(f"¡Perdiste! Tu nuevo saldo es ${user_data.wallet:,}.")

    async def get_user_data(self, modalidad, user_id):
        # Implementar la lógica para obtener los datos del usuario
        pass

    async def update_user_data(self, modalidad, user_data):
        # Implementar la lógica para actualizar los datos del usuario
        pass

async def setup(bot):
    await bot.add_cog(Gambling(bot))