Economy-Project/requirements.txt
discord.py
pygame
aiohttp
requests
python-dotenv
pytest
pytest-asyncio
Pillow
sqlalchemy
tenorpy